// src/app/sections/claim-mgt/claim/conversation/to-timeline-info/model/timeline-action.ts
export enum TimelineAction {
    partyModification = 'partyModification',
    changesAdded = 'changesAdded',
    filesAdded = 'filesAdded',
    commentAdded = 'commentAdded',
    statusReview = 'statusReview',
    statusPending = 'statusPending',
    statusDenied = 'statusDenied',
    statusRevoked = 'statusRevoked',
    statusAccepted = 'statusAccepted'
}